+++ models/settings/settings.go
@@ -63,6 +63,7 @@ type Settings struct {
 	InstagramAuth AuthProviderConfig `form:"instagramAuth" json:"instagramAuth"`
 	VKAuth        AuthProviderConfig `form:"vkAuth" json:"vkAuth"`
 	YandexAuth    AuthProviderConfig `form:"yandexAuth" json:"yandexAuth"`
+	PatreonAuth   AuthProviderConfig `form:"patreonAuth" json:"patreonAuth"`
 }
 
 // New creates and returns a new default Settings instance.
@@ -187,6 +188,9 @@ func New() *Settings {
 		YandexAuth: AuthProviderConfig{
 			Enabled: false,
 		},
+		PatreonAuth: AuthProviderConfig{
+			Enabled: false,
+		},
 	}
 }
 
@@ -230,6 +234,7 @@ func (s *Settings) Validate() error {
 		validation.Field(&s.InstagramAuth),
 		validation.Field(&s.VKAuth),
 		validation.Field(&s.YandexAuth),
+		validation.Field(&s.PatreonAuth),
 	)
 }
 
@@ -296,6 +301,7 @@ func (s *Settings) RedactClone() (*Settings, error) {
 		&clone.InstagramAuth.ClientSecret,
 		&clone.VKAuth.ClientSecret,
 		&clone.YandexAuth.ClientSecret,
+		&clone.PatreonAuth.ClientSecret,
 	}
 
 	// mask all sensitive fields
@@ -336,6 +342,7 @@ func (s *Settings) NamedAuthProviderConfigs() map[string]AuthProviderConfig {
 		auth.NameInstagram:  s.InstagramAuth,
 		auth.NameVK:         s.VKAuth,
 		auth.NameYandex:     s.YandexAuth,
+		auth.NamePatreon:    s.PatreonAuth,
 	}
 }