+++ models/settings/settings_test.go
@@ -73,6 +73,8 @@ func TestSettingsValidate(t *testing.T) {
 	s.VKAuth.ClientId = ""
 	s.YandexAuth.Enabled = true
 	s.YandexAuth.ClientId = ""
+	s.PatreonAuth.Enabled = true
+	s.PatreonAuth.ClientId = ""
 
 	// check if Validate() is triggering the members validate methods.
 	err := s.Validate()
@@ -114,6 +116,7 @@ func TestSettingsValidate(t *testing.T) {
 		`"instagramAuth":{`,
 		`"vkAuth":{`,
 		`"yandexAuth":{`,
+		`"patreonAuth":{`,
 	}
 
 	errBytes, _ := json.Marshal(err)
@@ -187,6 +190,8 @@ func TestSettingsMerge(t *testing.T) {
 	s2.VKAuth.ClientId = "vk_test"
 	s2.YandexAuth.Enabled = true
 	s2.YandexAuth.ClientId = "yandex_test"
+	s2.PatreonAuth.Enabled = true
+	s2.PatreonAuth.ClientId = "patreon_test"
 
 	if err := s1.Merge(s2); err != nil {
 		t.Fatal(err)
@@ -277,6 +282,7 @@ func TestSettingsRedactClone(t *testing.T) {
 	s1.InstagramAuth.ClientSecret = testSecret
 	s1.VKAuth.ClientSecret = testSecret
 	s1.YandexAuth.ClientSecret = testSecret
+	s1.PatreonAuth.ClientSecret = testSecret
 
 	s1Bytes, err := json.Marshal(s1)
 	if err != nil {
@@ -335,6 +341,7 @@ func TestNamedAuthProviderConfigs(t *testing.T) {
 	s.InstagramAuth.ClientId = "instagram_test"
 	s.VKAuth.ClientId = "vk_test"
 	s.YandexAuth.ClientId = "yandex_test"
+	s.PatreonAuth.ClientId = "patreon_test"
 
 	result := s.NamedAuthProviderConfigs()
 
@@ -366,6 +373,7 @@ func TestNamedAuthProviderConfigs(t *testing.T) {
 		`"instagram":{"enabled":false,"clientId":"instagram_test"`,
 		`"vk":{"enabled":false,"clientId":"vk_test"`,
 		`"yandex":{"enabled":false,"clientId":"yandex_test"`,
+		`"patreon":{"enabled":false,"clientId":"patreon_test"`,
 	}
 	for _, p := range expectedParts {
 		if !strings.Contains(encodedStr, p) {