+++ tools/auth/auth_test.go
@@ -207,4 +207,13 @@ func TestNewProviderByName(t *testing.T) {
 	if _, ok := p.(*auth.Yandex); !ok {
 		t.Error("Expected to be instance of *auth.yandex")
 	}
+
+	// patreon
+	p, err = auth.NewProviderByName(auth.NamePatreon)
+	if err != nil {
+		t.Errorf("Expected nil, got error %v", err)
+	}
+	if _, ok := p.(*auth.Patreon); !ok {
+		t.Error("Expected to be instance of *auth.Patreon")
+	}
 }